mamacare/requirements.txt
# Core Dependencies
langchain>=0.1.0
langgraph>=0.0.10
google-generativeai>=0.3.0

# Web Framework
Flask>=2.3.3
Flask-CORS>=4.0.0
Werkzeug>=2.3.7
Jinja2>=3.1.2
itsdangerous>=2.1.2
click>=8.1.7

# Utilities
python-dotenv>=1.0.0
apscheduler>=3.10.1
python-dateutil>=2.8.2
pytz>=2023.3

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.1

# Documentation
Sphinx>=7.0.1
sphinx-rtd-theme>=1.2.2

# Production
waitress>=2.1.2
gunicorn>=21.2.0
